Refactor the tmux panes to display more useful information
This will remove any noise from the prompt by removing kubernetes
information and can better utilise the tmux panes. This could
potentialyl be a better fit. Also adding the status of the last exit
code, so had to disable a functionality in fasd that was adding a hook
in the PROMPT_COMMAND.